updated lm and mp fitting tolerance to match. Upped num of iter for int spec fitting from 1000 to 2000

+++ src/fitting/optimizers/optimizer.h
@@ -121,7 +121,7 @@ void fill_user_data(User_Data &ud,
                     const Base_Model * const model,
                     const Range energy_range,
                     Callback_Func_Status_Def* status_callback,
-                    int total_itr,
+                    size_t total_itr,
                     bool use_weights = false);
 
 void fill_gen_user_data(Gen_User_Data &ud,
